Rename on(), off() functions to watch(), ignore()
I've only renamed the fields on the returned singleton for now to keep things simple, but I think `on()` and `off()` functions don't make sense without an attributed action. Typically you'd get `on('click',...)` for example.

This just makes things clear that you're setting up a watcher/listener on the CSS selector.